+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Background color selection
+- Settings are saved to file
 
 ## [1.0.0] - 2022-02-27
 - Initial version